# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiseAuVert API",
        "description": "API documentation for MiseAuVert",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://app.miseauvert.fr",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/pensions": {
            "get": {
                "tags": [
                    "Pensions"
                ],
                "summary": "Récupérer toutes les pensions",
                "operationId": "cf327193d22f43783ec6d0309c4a745e",
                "responses": {
                    "200": {
                        "description": "Liste des pensions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pension"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pensions"
                ],
                "summary": "Créer une pension",
                "operationId": "d77907a44ef52eb9c5b3ba77ff88743e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "adresse",
                                    "ville",
                                    "code_postal",
                                    "actif"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "region": {
                                        "type": "string"
                                    },
                                    "code_postal": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "capacite_chiens": {
                                        "type": "integer"
                                    },
                                    "capacite_chats": {
                                        "type": "integer"
                                    },
                                    "actif": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pension créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pension"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/pensions/{id}": {
            "get": {
                "tags": [
                    "Pensions"
                ],
                "summary": "Afficher une pension",
                "operationId": "96f84d56fb0e4e83b8edc2c6cd1838a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pension trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pension"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pensions"
                ],
                "summary": "Mettre à jour une pension",
                "operationId": "1e4a2f8dcaf9b464f177356dfc1adcfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pension"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pension mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pension"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pensions"
                ],
                "summary": "Supprimer une pension",
                "operationId": "b676ef9735e1adaa9201df31a04699fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pension supprimée"
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    }
                }
            }
        },
        "/api/pensions/{id}/boxes": {
            "get": {
                "tags": [
                    "Boxes"
                ],
                "summary": "Récupérer les boxes d'une pension",
                "operationId": "d78121ff4d85a1488e3ac29eceaacc38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Boxes récupérées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Box"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    }
                }
            },
            "post": {
                "tags": [
                    "Boxes"
                ],
                "summary": "Créer une box",
                "operationId": "ac24f51f02b6236d86f7f31b3e60f25b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "superficie": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Box créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Box"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/boxes/{id}": {
            "put": {
                "tags": [
                    "Boxes"
                ],
                "summary": "Mettre à jour une box",
                "operationId": "655fd14e8f1f3412c0b0a0da60cea6e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "superficie": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Box mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Box"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Box non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Boxes"
                ],
                "summary": "Supprimer une box",
                "operationId": "909871f34a0e87678974c22badb5783e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Box supprimée"
                    },
                    "404": {
                        "description": "Box non trouvée"
                    }
                }
            }
        },
        "/api/pensions/{id}/types-gardiennage": {
            "get": {
                "tags": [
                    "Types de Gardiennage"
                ],
                "summary": "Récupérer les types de gardiennage d'une pension",
                "operationId": "4b2720f1594625d5ca71cdf998029516",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Types de gardiennage récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TypeGardiennage"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    }
                }
            },
            "post": {
                "tags": [
                    "Types de Gardiennage"
                ],
                "summary": "Créer un type de gardiennage",
                "operationId": "93ad189f221467d8be341558f76c3b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type de gardiennage créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TypeGardiennage"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pension non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Pension": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "ville": {
                        "type": "string"
                    },
                    "adresse": {
                        "type": "string"
                    },
                    "telephone": {
                        "type": "string"
                    },
                    "responsable": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Box": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "pension_id": {
                        "type": "integer"
                    },
                    "superficie": {
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez le token obtenu via /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}